Add boolean (no param) solver
